requirements.txt
Django==4.2.7
paramiko==3.3.1
python-dotenv==1.0.0
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0
